Static Linking
Added static linking! Expressed as follows:
- Adding the "import" keyword before an int variable will mark it to be
imported on linking
- Behind the scenes an imported variable contains an address relative to
the variable's location which points the real address
- Reading from imported vars is supported (retrieves an absolute address
to the imported variable) as well as the & tag which will resolve the
pointer to an int.
- You can write to a variable but at the risk of corrupting the pointer.
It requires a free register (and r0) to write properly:
go -> r1
L@int[r1] <- r2
- Reading to r0 from a import var is not supported: r0 is used as a
temporary register for holding the var pointer
- Variables with the extern keyword can be imported upon linking, their
behavior is not changed
There's an example on how to use linking, TestAsm and LinkAsm.
Added the --link command line option to link files -> --link outFile
masterFile linkCount linkFiles{linkCount}
Added a required argument to --assemble, the assemble type. ACCEDEFF
creates linkable files, FLAT_FILE is like legacy.